+++ QuizWebsite/src/main/java/org/ja/dao/ChallengesDao.java
@@ -2,6 +2,7 @@ package org.ja.dao;
 
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.ja.model.OtherObjects.Challenge;
+import org.ja.model.OtherObjects.Message;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -22,12 +23,15 @@ create table challenges(
  */
 public class ChallengesDao {
     private final BasicDataSource dataSource;
-
+    private long cnt=0;
     public ChallengesDao(BasicDataSource dataSource) {
         this.dataSource = dataSource;
     }
 
     public void insertChallenge(Challenge challenge) {
+        if(contains(challenge)){
+            return;
+        }
         String sql = "INSERT INTO challenges (sender_user_id, recipient_user_id, quiz_id) VALUES (?,?,?)";
 
         try(Connection c = dataSource.getConnection();
@@ -40,8 +44,10 @@ public class ChallengesDao {
             ps.executeUpdate();
 
             try (ResultSet rs = ps.getGeneratedKeys()){
-                if (rs.next())
+                if (rs.next()) {
+                    cnt++;
                     challenge.setChallengeId(rs.getLong("challenge_id"));
+                }
             }
 
         } catch (SQLException e) {
@@ -50,6 +56,9 @@ public class ChallengesDao {
     }
 
     public void removeChallenge(long challengeId) {
+        if(!contains(challengeId)){
+            return;
+        }
         String sql = "DELETE FROM challenges WHERE challenge_id = ?";
 
         try (Connection c= dataSource.getConnection();
@@ -58,6 +67,7 @@ public class ChallengesDao {
             ps.setLong(1, challengeId);
 
             ps.executeUpdate();
+            cnt--;
         } catch (SQLException e) {
             throw new RuntimeException("Error removing challenge from database", e);
         }
@@ -106,7 +116,54 @@ public class ChallengesDao {
 
         return challenges;
     }
+    public boolean contains(Challenge c){
+        if(c==null){
+            return false;
+        }
+        String sql = "SELECT COUNT(*) FROM challenges WHERE  challenge_id=? AND sender_user_id=?" +
+                "AND recipient_user_id=? AND quiz_id=?";
 
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+            ps.setLong(1, c.getChallengeId());
+            ps.setLong(2, c.getSenderUserId());
+            ps.setLong(3, c.getRecipientUserId());
+            ps.setLong(4, c.getQuizId());
+            ResultSet rs = ps.executeQuery();
+
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+
+            return false;
+        } catch (SQLException e) {
+            throw new RuntimeException("Error checking user existence", e);
+        }
+    }
+    public boolean contains(long cid){
+        if(cid<0||cid>cnt){
+            return false;
+        }
+        String sql = "SELECT COUNT(*) FROM challenges WHERE challenge_id = ?";
+
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+
+            ps.setLong(1, cid);
+            ResultSet rs = ps.executeQuery();
+
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+
+            return false;
+        } catch (SQLException e) {
+            throw new RuntimeException("Error checking user existence", e);
+        }
+    }
+    public long getCount(){
+        return cnt;
+    }
     private Challenge retrieveChallenge(ResultSet rs) throws SQLException {
         return new Challenge(rs.getLong("challenge_id"), rs.getLong("sender_user_id"),
                 rs.getLong("recipient_user_id"), rs.getLong("quiz_id"));
