+++ QuizWebsite/src/main/java/org/ja/dao/TagsDao.java
@@ -14,12 +14,15 @@ import java.util.ArrayList;
 
 public class TagsDao {
     private final BasicDataSource dataSource;
-
+    private long cnt=0;
     public TagsDao(BasicDataSource dataSource) {
         this.dataSource = dataSource;
     }
 
-    public void addTag(Tag tag) {
+    public void insertTag(Tag tag) {
+        if(containsTag(tag.getTagName())) {
+            return;
+        }
         String sql = "INSERT INTO tags (tag_name) VALUES (?)";
 
         try (Connection c = dataSource.getConnection();
@@ -32,6 +35,7 @@ public class TagsDao {
 
             try (ResultSet keys = preparedStatement.getGeneratedKeys()){
                 if (keys.next()) {
+                    cnt++;
                     long newId = keys.getLong(1);
                     tag.setTagId(newId); // if you want to store it in your object
                 }
@@ -42,6 +46,9 @@ public class TagsDao {
     }
 
     public void removeTag(Tag tag) {
+        if(!containsTag(tag.getTagName())) {
+            return;
+        }
         String sql = "DELETE FROM tags WHERE tag_id = ?";
         try (Connection c = dataSource.getConnection();
             PreparedStatement preparedStatement =
@@ -49,6 +56,7 @@ public class TagsDao {
 
             preparedStatement.setLong(1, tag.getTagId());
             preparedStatement.executeUpdate();
+            cnt--;
         } catch (SQLException e) {
             throw new RuntimeException("Error removing tag from database", e);
         }
@@ -88,6 +96,45 @@ public class TagsDao {
             throw new RuntimeException("Error querying tag from by name database", e);
         }
         return null;
+    }
+    public boolean containsTag(String name) {
+        String sql = "SELECT COUNT(*) FROM tags WHERE tag_name = ?";
+
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+
+            ps.setString(1, name);
+            ResultSet rs = ps.executeQuery();
+
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+
+            return false;
+        } catch (SQLException e) {
+            throw new RuntimeException("Error checking user existence", e);
+        }
+    }
+    public boolean containsTag(long id) {
+        String sql = "SELECT COUNT(*) FROM tags WHERE tag_id = ?";
+
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+
+            ps.setLong(1, id);
+            ResultSet rs = ps.executeQuery();
+
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+
+            return false;
+        } catch (SQLException e) {
+            throw new RuntimeException("Error checking user existence", e);
+        }
+    }
+    public long getCount() {
+        return cnt;
     }
     /*public ArrayList<Tag> filterCategories(Filter filter) {
         String sql="SELECT * FROM categories WHERE "+filter.toString();
