+++ QuizWebsite/src/main/java/org/ja/dao/UsersDao.java
@@ -12,14 +12,17 @@ import java.util.ArrayList;
 
 public class UsersDao {
     private final BasicDataSource dataSource;
-
+    private long cnt=0;
     public UsersDao(BasicDataSource dataSource) {
         this.dataSource = dataSource;
     }
 
     public void insertUser(User user) throws SQLException {
+        if(containsUser(user.getUsername())){
+            return;
+        }
         String sql = "INSERT INTO users (password_hashed, username, user_photo, user_status, salt) " +
-                "VALUES (?, ?, ?, ?, ?, ?);";
+                "VALUES (?, ?, ?, ?, ?);";
         try (Connection c = dataSource.getConnection();
             PreparedStatement preparedStatement =
                     c.prepareStatement(sql, PreparedStatement.RETURN_GENERATED_KEYS)){
@@ -34,6 +37,7 @@ public class UsersDao {
 
             try (ResultSet keys = preparedStatement.getGeneratedKeys()){
                 if (keys.next()) {
+                    cnt++;
                     long newId = keys.getLong(1);
                     user.setId(newId); // if you want to store it in your object
                 }
@@ -44,6 +48,9 @@ public class UsersDao {
     }
 
     public void removeUserById(long id) {
+        if(!containsUser(id)){
+            return;
+        }
         String sql = "DELETE FROM users WHERE user_id = ?";
 
         try (Connection c = dataSource.getConnection();
@@ -51,12 +58,16 @@ public class UsersDao {
             preparedStatement.setLong(1, id);
 
             preparedStatement.executeUpdate();
+            cnt--;
         } catch (SQLException e) {
             throw new RuntimeException("Error removing user by id from database", e);
         }
     }
 
     public void removeUserByName(String name) {
+        if(!containsUser(name)){
+            return;
+        }
         String sql = "DELETE FROM users WHERE username = ?";
 
         try(Connection c=dataSource.getConnection();
@@ -65,6 +76,7 @@ public class UsersDao {
             preparedStatement.setString(1, name);
 
             preparedStatement.executeUpdate();
+            cnt--;
         } catch (SQLException e) {
             throw new RuntimeException("Error removing user by name from database", e);
         }
@@ -124,7 +136,46 @@ public class UsersDao {
 
         return users;
     }
+    public boolean containsUser(String username) {
+        String sql = "SELECT COUNT(*) FROM users WHERE username = ?";
 
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+
+            ps.setString(1, username);
+            ResultSet rs = ps.executeQuery();
+
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+
+            return false;
+        } catch (SQLException e) {
+            throw new RuntimeException("Error checking user existence", e);
+        }
+    }
+    public boolean containsUser(long id) {
+        String sql = "SELECT COUNT(*) FROM users WHERE user_id = ?";
+
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+
+            ps.setLong(1, id);
+            ResultSet rs = ps.executeQuery();
+
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+
+            return false;
+        } catch (SQLException e) {
+            throw new RuntimeException("Error checking user existence", e);
+        }
+    }
+
+    public long getCount(){
+        return cnt;
+    }
     private User retrieveUser(ResultSet rs) throws SQLException {
         return new User(rs.getLong("user_id"), rs.getString("username"),
                 rs.getString("password_hashed"), rs.getString("salt"),
